requirements.txt
flask
flask-cors
tensorflow
keras
pillow
numpy
gunicorn